+++ api/pokerApi.js
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const pokerApi = axios.create({
+  baseURL: "https://bicyclecards.com/how-to-play/",
+});