+++ api/wallpapersURL.js
@@ -2,5 +2,4 @@ import axios from "axios";
 
 export const wallpapersURL = axios.create({
   baseURL: "https://sol-moon-wallpapers.vercel.app/",
-  responseType: "stream"
 });