+++ hash/utilityTools.js
@@ -1,7 +1,11 @@
+// menu
+import { stackMenu } from "../menu.js";
+import { pokerGameOpts } from "../utils.js";
+
 // functions
 import genPassword from "../functions/password.js";
 import hardware from "../functions/hardware.js";
-
+import pokerGame from "../functions/poker.js";
 
 const utilityTools = {
   password(refreshCallback) {
@@ -13,6 +17,15 @@ const utilityTools = {
     console.clear();
     hardware();
     setTimeout(refreshCallback, 3e3);
+  },
+  async poker_game(refreshCallback) {
+  	const pokeOpt = await stackMenu({
+  		message: "select a poker game for view instructions:",
+  		choices: pokerGameOpts
+  	});
+  	
+  	pokerGame(pokeOpt);
+  	setTimeout(refreshCallback, 5e3);
   }
 };