+++ hash/wallpaperSelect.js
@@ -1,6 +1,5 @@
 // inquirer
-import inquirer from "inquirer";
-import colors from "colors";
+import { select } from "@inquirer/prompts";
 
 import { wallpaperDownload } from "../functions/download.js";
 
@@ -8,68 +7,52 @@ const message = "select a wallpaper for download:";
 const backMenu = "back to menu";
 
 const wallpaperSelect = {
-  solMoon(refreshCallback, alternativeCallback) {
+  async solMoon(refreshCallback, alternativeCallback) {
     const solMoonWallpapers = [...Array(20).keys()]
       .map(i => `sol-moon${i + 1}.jpeg`);	
 
-    inquirer.prompt([
-      {
-        type: "list",
-        name: "solMoon",
-        message,
-        choices: [...solMoonWallpapers, backMenu]
-      }
-    ])
-      .then(({ solMoon }) => {
-        if (solMoon === backMenu) {
-          alternativeCallback();
-        } else {
-          wallpaperDownload("sol-moon", solMoon);
-          setTimeout(refreshCallback, 5000);
-        }
-      })
-      .catch(err => console.error(colors.red(err.message)));
+    
+    const solMoon = await select({
+    	message, choices: [...solMoonWallpapers, backMenu]
+    });
+    
+    if (solMoon === backMenu) {
+    	alternativeCallback();
+    } else {
+    	wallpaperDownload("sol-moon", solMoon);
+    	setTimeout(refreshCallback, 5000);
+    }
   },
-  dimensions(refreshCallback, alternativeCallback) {
+  async dimensions(refreshCallback, alternativeCallback) {
     const dimensionsWallpapers = [...Array(12).keys()]
       .map(i => `dimensions-${i + 1}.jpeg`);
 
-    inquirer.prompt([
-      {
-        type: "list",
-        name: "dimensions",
-        message,
-        choices: [...dimensionsWallpapers, backMenu]
-      }
-    ])
-      .then(({ dimensions }) => {
-        if(dimensions === backMenu) {
-          alternativeCallback();    
-        } else {
-          wallpaperDownload("dimensions", dimensions);
-          setTimeout(refreshCallback, 5000);
-        }
-      });
+    
+    const dimensions = await select({
+    	message, choices: [...dimensionsWallpapers, backMenu]
+    });
+    
+    if(dimensions === backMenu) {
+    	alternativeCallback();    
+    } else {
+    	wallpaperDownload("dimensions", dimensions);
+    	setTimeout(refreshCallback, 5000);
+    }
   },
-  seyyahi2(refreshCallback, alternativeCallback) {
+  async seyyahi2(refreshCallback, alternativeCallback) {
   	const seyyahiWallpapers = [...Array(14).keys()]
   		.map(i => `seyyahi2-wallpaper${i + 1}.jpg`);
-  		
-  	inquirer.prompt([
-  		{
-  			type: "list",
-  			name: "seyyahi2",
-  			message,
-  			choices: [...seyyahiWallpapers, backMenu]
-  		}
-  	])
-  		.then(({ seyyahi2 }) => {
-  			if(seyyahi2 === backMenu) {
-  				alternativeCallback();
-  			} else {
-  				wallpaperDownload("seyyahi2", seyyahi2);
-  			}
-  		});
+  	
+  	const seyyahi2 = await select({
+  		message, choices: [...seyyahiWallpapers, backMenu]
+  	});
+  	
+  	if(seyyahi2 === backMenu) {
+  		alternativeCallback();
+  	} else {
+  		wallpaperDownload("seyyahi2", seyyahi2);
+  		setTimeout(refreshCallback, 5000);
+  	}
   }
 };