+++ hash/webTools.js
@@ -2,7 +2,7 @@
 import { performance } from "node:perf_hooks";
 
 // inquirer
-import inquirer from "inquirer";
+import { input } from "@inquirer/prompts";
 
 // functions
 import singleStack from "../functions/singleStack.js";
@@ -11,57 +11,81 @@ import pageSpeed from "../functions/pageSpeed.js";
 import scrape from "../functions/scraping.js";
 import cssValidate from "../functions/cssValidator.js";
 
-// validations
-import { 
-  multipleWebQuery, 
-  singleWebQuery, 
-  webScrapingQuery 
-} from "../validations/webValidations.js";
+import { stackMenu } from "../menu.js";
+import { scrapingOpts } from "../utils.js";
+
+// regex
+const webRegex = /https?:\/\//g;
 
 const webTools = {
-  single(refreshCallback) {
+  async single(refreshCallback) {
     console.clear();
-    inquirer.prompt([singleWebQuery])
-      .then(({ url }) => {
-        singleStack(url);
-        const timeEnd = performance.now();
-        setTimeout(refreshCallback, timeEnd);
-      });
+    
+    const url = await input({
+    	message: "enter a url:",
+    	validate: input => webRegex.test(input) || "enter a url valid".yellow
+    });  
+      
+    singleStack(url);
+    const timeEnd = performance.now();
+    setTimeout(refreshCallback, timeEnd);
   },
-  multiple(refreshCallback) {
+  async multiple(refreshCallback) {
     console.clear();
-    inquirer.prompt([multipleWebQuery])
-      .then(({webList}) => {
-        multipleStack(webList.split(" "));
-        const timeEnd = performance.now();
-        setTimeout(refreshCallback, timeEnd);
-      });
+    
+    const webList = await input({
+    	message: "enter URLs for analyze the tech stacks with whitespace without quotes example 'http://example.com https://nodejs.org': \n",
+    	validate(input) {
+    		const pass = input.match(webRegex);
+
+    		return pass && pass.length === 2 || "must be 2 sites";
+    	}
+    });
+    
+    multipleStack(webList.split(" "));
+    const timeEnd = performance.now();
+    setTimeout(refreshCallback, timeEnd);
   },
-  pagespeed(refreshCallback) {
+  async pagespeed(refreshCallback) {
     console.clear();
-    inquirer.prompt([singleWebQuery])
-      .then(({ url }) => {
-        pageSpeed(url);
-        const timeEnd = performance.now();
-        setTimeout(refreshCallback, timeEnd);
-      });
+    
+    const url = await input({
+    	message: "enter a url:",
+    	validate: input => webRegex.test(input) || "enter a url valid".yellow
+    });
+    
+    pageSpeed(url);
+    const timeEnd = performance.now();
+    setTimeout(refreshCallback, timeEnd);
   },
-  scraping(refreshCallback) {
+  async scraping(refreshCallback) {
     console.clear();
-    inquirer.prompt([singleWebQuery, webScrapingQuery])
-      .then(({ url, option }) => {
-        scrape(url, option);
-        setTimeout(refreshCallback, 3000);
-      });
+    
+    const { url, option } = {
+      option: await stackMenu({
+        pageSize: 9,
+        message: "select a web scraping option:",
+        choices: scrapingOpts
+      }),
+      url: await input({
+        message: "enter a url:",
+        validate: input => webRegex.test(input) || "enter a url valid".yellow
+      })
+    };
+      
+    scrape(url, option);
+    setTimeout(refreshCallback, 3000);
   },
-  css_validate(refreshCallback) {
+  async css_validate(refreshCallback) {
     console.clear();
-    inquirer.prompt([singleWebQuery])
-      .then(({ url }) => {
-        cssValidate(url);
-        const timeEnd = performance.now();
-        setTimeout(refreshCallback, timeEnd);
-      });
+    const url = await input({
+    	message: "enter a url:",
+    	validate: input => webRegex.test(input) || "enter a url valid".yellow
+    });
+      
+    cssValidate(url);
+    const timeEnd = performance.now();
+    setTimeout(refreshCallback, timeEnd);
   }
 };